Moved TargetDevices.java to other package
