+++ src/main/java/com/odeyalo/sonata/connect/service/player/DefaultDeviceOperations.java
@@ -4,8 +4,8 @@ import com.odeyalo.sonata.connect.model.CurrentPlayerState;
 import com.odeyalo.sonata.connect.model.Device;
 import com.odeyalo.sonata.connect.model.Devices;
 import com.odeyalo.sonata.connect.model.User;
+import com.odeyalo.sonata.connect.service.TargetDevices;
 import com.odeyalo.sonata.connect.service.player.handler.TransferPlaybackCommandHandlerDelegate;
-import com.odeyalo.sonata.connect.service.player.sync.TargetDevices;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
