+++ src/main/java/com/odeyalo/sonata/connect/service/player/EventPublisherDeviceOperationsDecorator.java
@@ -4,8 +4,8 @@ import com.odeyalo.sonata.connect.model.CurrentPlayerState;
 import com.odeyalo.sonata.connect.model.Device;
 import com.odeyalo.sonata.connect.model.Devices;
 import com.odeyalo.sonata.connect.model.User;
+import com.odeyalo.sonata.connect.service.TargetDevices;
 import com.odeyalo.sonata.connect.service.player.sync.PlayerSynchronizationManager;
-import com.odeyalo.sonata.connect.service.player.sync.TargetDevices;
 import com.odeyalo.sonata.connect.service.player.sync.event.DeviceConnectedPlayerEvent;
 import com.odeyalo.sonata.connect.service.player.sync.event.DeviceDisconnectedPlayerEvent;
 import org.jetbrains.annotations.NotNull;
